Post Type Fix
Added line 196: If I was on a singular entry of a post type, the object
returned by $wp_query->get_queried_object was not the object for the
post type, but the post object for the singular entry. Using
get_post_type_object makes sure that you have the correct object before
going forward on singular entries.